Address review

+++ disk/smart-monitoring.sh
@@ -21,7 +21,7 @@ debug_mode
 root_check
 
 # Check if bpytop is already installed
-if ! [ -f "$SCRIPTS/smart-notification.sh" ]
+if ! is_this_installed smartmontools
 then
     # Ask for installing
     install_popup "$SCRIPT_NAME"
@@ -29,27 +29,86 @@ else
     # Ask for removal or reinstallation
     reinstall_remove_menu "$SCRIPT_NAME"
     # Removal
-    check_command rm "$SCRIPTS/smart-notification.sh"
-    if is_this_installed smartmontools
-    then
-        apt purge smartmontools -y
-        apt autoremove -y
-    fi
+    rm -f "$SCRIPTS/smart-notification.sh"
+    check_command apt purge smartmontools -y
+    apt autoremove -y
+    # reset the cronjob
+    crontab -u root -l | grep -v 'smartctl.sh'  | crontab -u root -
+    crontab -u root -l | grep -v 'smart-notification.sh'  | crontab -u root -
     # Show successful uninstall if applicable
     removal_popup "$SCRIPT_NAME"
+fi
+
+# Choose between direct notifiaction or weekly
+choice=$(whiptail --title "$TITLE" --menu \
+"Please choose if you want to get informed weekly or directly if an error occurs.
+$MENU_GUIDE\n\n$RUN_LATER_GUIDE" "$WT_HEIGHT" "$WT_WIDTH" 4 \
+"Directly" "(Continous S.M.A.R.T checking)" \
+"Weekly" "(Weekly S.M.A.R.T checking)" 3>&1 1>&2 2>&3)
+
+# Exit if nothing chosen
+if [ -z "$choice" ]
+then
+    exit 1
 fi
 
 # Install needed tools
 install_if_not smartmontools
 check_command systemctl stop smartmontools
 
-# Write conf to file
-# https://wiki.debianforum.de/Festplattendiagnostik-_und_%C3%9Cberwachung#Beispiel_3
-echo "DEVICESCAN -a -I 194 -W 5,45,55 -r 5 -R 5 -n standby,24 -m <nomailer> -M exec $SCRIPTS/smart-notification.sh \
--s (O/../.././(06|18)|S/../../6/02|L/../20/./02)" > /etc/smartd.conf
+# Weekly notification
+if [ "$choice" = "Weekly" ]
+then
+    # Create smart notification script
+    cat << SMART_NOTIFICATION > "$SCRIPTS/smart-notification.sh"
+#!/bin/bash
 
-# Create smart notification script
-cat << SMART_NOTIFICATION > "$SCRIPTS/smart-notification.sh"
+# T&M Hansson IT AB © - 2020, https://www.hanssonit.se/
+# Copyright © 2020 Simon Lindner (https://github.com/szaimen)
+
+# shellcheck disable=2034,2059
+true
+SCRIPT_NAME="S.M.A.R.T Notification"
+SCRIPT_EXPLAINER="This script sends a notification if something is wrong with your drives."
+
+# shellcheck source=lib.sh
+source /var/scripts/fetch_lib.sh || source <(curl -sL https://raw.githubusercontent.com/nextcloud/vm/master/lib.sh)
+
+# Check if root
+root_check
+if home_sme_server
+then
+    notify_admin_gui "S.M.A.R.T results weekly scan (nvme0n1)" "$(smartctl --all /dev/nvme0n1)"
+    notify_admin_gui "S.M.A.R.T results weekly scan (sda)" "$(smartctl --all /dev/sda)"
+else
+    # get all disks into an array
+    disks="$(fdisk -l | grep Disk | grep /dev/sd | awk '{print$2}' | cut -d ":" -f1)"
+    # loop over disks in array
+    for disk in $(printf "${disks[@]}")
+    do
+        if [ -n "$disks" ]
+        then
+             notify_admin_gui "S.M.A.R.T results weekly scan ($disk)" "$(smartctl --all $disk)"
+        fi
+    done
+fi
+SMART_NOTIFICATION
+    # Add crontab “At 06:12 on Monday.”
+    if ! crontab -u root -l | grep -w 'smart-notification.sh'
+    then
+        print_text_in_color "$ICyan" "Adding weekly crontab..."
+        crontab -u root -l | { cat; echo "12 06 * * 1 $SCRIPTS/smart-notification.sh"; } | crontab -u root -
+    fi
+# Direct notification
+elif [ "$choice" = "Directly" ]
+then
+    # Write conf to file
+    # https://wiki.debianforum.de/Festplattendiagnostik-_und_%C3%9Cberwachung#Beispiel_3
+    echo "DEVICESCAN -a -I 194 -W 5,45,55 -r 5 -R 5 -n standby,24 -m <nomailer> -M exec \
+$SCRIPTS/smart-notification.sh -s (O/../../(2|4|6)/02|S/../../5/02|L/../20/./02)" > /etc/smartd.conf
+
+    # Create smart notification script
+    cat << SMART_NOTIFICATION > "$SCRIPTS/smart-notification.sh"
 #!/bin/bash
 
 # T&M Hansson IT AB © - 2020, https://www.hanssonit.se/
@@ -79,9 +138,11 @@ then
 fi
 exit
 SMART_NOTIFICATION
+fi
 
 # Make it executable
-chmod +x "$SCRIPTS/smart-notification.sh"
+chown root:root "$SCRIPTS/smart-notification.sh"
+chmod 700 "$SCRIPTS/smart-notification.sh"
 
 # Restart service
 if start_if_stopped smartmontools
